design doc and more refactoring
